Tarea1SD/cache_memcach/app/search/requirements.txt
grpcio==1.57.0
grpcio-tools==1.57.0
uhashring==2.3
numpy==1.21.5
python-memcached==1.59

